requirements.txt
Flask==2.1.2
gunicorn==20.1.0
line-bot-sdk==1.20.0
openai==0.27.0